+++ backend/app/schemas/payroll.py
@@ -1,14 +1,15 @@
 from pydantic import BaseModel
 from typing import Optional
+from decimal import Decimal
 
 class PayrollOut(BaseModel):
     id: int
     employee_id: int
     month: str
-    basic: float
-    hra: float
-    allowances: float
-    deductions: float
-    net_salary: float
-    tax_deducted: float
+    basic: Decimal
+    hra: Decimal
+    allowances: Decimal
+    deductions: Decimal
+    net_salary: Decimal
+    tax_deducted: Decimal
     model_config = {"from_attributes": True}
